+++ README.md
@@ -1,5 +1,25 @@
-# Файлы для итогового задания
+# Планировщик задач
 
-В директории `tests` находятся тесты для проверки API, которое должно быть реализовано в веб-сервере.
+> Планировщик хранит задачи, каждая из них содержит дату дедлайна и заголовок с комментарием. 
+> Задачи могут повторяться по заданному правилу: например, ежегодно, через какое-то количество дней, в определённые дни месяца или недели. 
+> Если отметить такую задачу как выполненную, она переносится на следующую дату в соответствии с правилом. Обычные задачи при выполнении будут просто удаляться.
 
-Директория `web` содержит файлы фронтенда.
+## Запуск проекта и автотестов
+
+Запустить проект можно двумя способами:
+
+```
+go run .
+```
+
+```
+go build -o app
+./app
+```
+
+
+В директории `tests` находятся тесты для проверки API.
+
+```
+go test ./tests
+```
